multi-node-gpu-simulated/requirements.txt
tensorflow-gpu==2.10
typer
